backend/requirements.txt
fastapi>=0.104,<1.0
uvicorn>=0.22,<1.0
psycopg2-binary>=2.9,<3.0

